move InfoBox to a separate file as its own component
